day1: reorder add instruction
There is a data hazard between

  lw x11,0(x2)

And

  add x12,x10,x11

The pipelined implementation inserts a bubble to resolve this dependency, as the
data is not ready when the add instruction is in the execute stage.

Reording this so the x2 (index variable) is incremented after the load removes
the need for the bubble insertion. This allows the value to be forwarded.